refactor(tests): remove `common` folder
