+++ tests/main_tests/basic_daemon_rpc.rs
@@ -1,6 +1,6 @@
 use std::str::FromStr;
 
-use crate::common::helpers;
+use super::helpers;
 use monero::cryptonote::hash::Hash;
 use monero_rpc::{HashString, TransactionsResponse};
