+++ tests/main_tests/helpers/wallet.rs
